unit test for #839
